Withdraw Selection Screen once selection is made
The initial selection screen from which the user can choose 'Site',
'Dictionary' or 'Editor' is hidden once the selection is made. This
avoids a bug by which two Editor instances will interfere with each
other's keypress events, HTML servers and current working directories.

This is a quick patch, and should be removed in the future, when a
better bugfix is found.