Take advantage of Smeagol Markdown
A recent change to the basic Smeagol Markdown has been to use [d foo]
syntax to signify an html <div class="foo"> element. This change builds
on that, by changing how the ExternalGrammar linkadder works. This
linkadder depends on the specific format of the "part of speech" /
"definition" line of a dictionary entry.