Add checkboxes for expansion and contraction
In the past, the :hover and :active css selectors were used separately
to show or hide menus and "folding" divs. This lead to issues when
multiple input methods are available, e.g.: when using a mouse and
touchscreen simultaneously.

Now we use the 'checkbox hack', which required folding divs and menus to
be automatically given <label> and <input> tags.